Original OG ActiveGS kegs_3_0_242 base.

+++ jni/scc_macdriver.c
@@ -70,7 +70,7 @@ scc_serial_mac_change_params(int port)
 
 	scc_ptr = &(scc_stat[port]);
 
-	fd = (int)scc_ptr->host_handle;
+	fd = (intptr_t)scc_ptr->host_handle;
 	printf("scc_serial_mac_change_parms port: %d, fd: %d\n", port, fd);
 	if(fd <= 0) {
 		return;
@@ -140,7 +140,7 @@ scc_serial_mac_fill_readbuf(int port, int space_left, double dcycs)
 
 	scc_ptr = &(scc_stat[port]);
 
-	fd = (int)scc_ptr->host_handle;
+	fd = (intptr_t)scc_ptr->host_handle;
 	if(fd <= 0) {
 		return;
 	}
@@ -170,7 +170,7 @@ scc_serial_mac_empty_writebuf(int port)
 
 	scc_ptr = &(scc_stat[port]);
 
-	fd = (int)scc_ptr->host_handle;
+	fd = (intptr_t)scc_ptr->host_handle;
 	if(fd <= 0) {
 		return;
 	}
